+++ @stellar/design-system-website/docs/components/buttons/button.mdx
@@ -9,7 +9,11 @@ slug: /button
 ## Example
 
 ```tsx live
-<PreviewBlock componentName="Button" />
+<PreviewBlock componentName="Button">
+  <Button variant="primary" size="md">
+    Button
+  </Button>
+</PreviewBlock>
 ```
 
 ## Props
