+++ @stellar/design-system-website/src/theme/MDXComponents.js
@@ -1,9 +1,11 @@
 import MDXComponents from "@theme-original/MDXComponents";
 import { ComponentDescription } from "@site/src/components/ComponentDescription";
 import { ComponentProps } from "@site/src/components/ComponentProps";
+import { PreviewBlock } from "@site/src/components/PreviewBlock";
 
 export default {
   ...MDXComponents,
   ComponentDescription,
   ComponentProps,
+  PreviewBlock,
 };
